+++ Editor/Core/ProjectStateManagement/IServiceFlagRequest.cs
@@ -1,9 +1,10 @@
+using System.Collections.Generic;
 using Unity.Services.Core.Internal;
 
 namespace Unity.Services.Core.Editor
 {
     interface IServiceFlagRequest
     {
-        IAsyncOperation<IServiceFlags> FetchServiceFlags();
+        IAsyncOperation<IDictionary<string, bool>> FetchServiceFlags();
     }
 }